requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
face_recognition
Pillow
